requirements.txt
pyserial~=3.5
pip install mpremote
